finalizando teste da forma mais simples

+++ CopaDoMundoDeFilmes/models/Movie.cs
@@ -5,6 +5,6 @@
         public string Id { get; set; }
         public string Titulo { get; set; }
         public int Ano { get; set; }
-        public decimal Nota { get; set; }
+        public float Nota { get; set; }
     }
 }
